+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Features:
 
+- added full support for generic list and map types
 - added full serialization support for embedded events
 
 ## 0.10.1 (2018-04-05)